CppEditor: Clear out mis-detected parentheses
... from the syntactic highlighter.
For instance, in this piece of code:
    void myfunc()
    {
        const auto rawString = R"(abc
        de)fg"hij)";
    }
The closing parenthesis inside the raw string literal is mis-interpreted
by SimpleLexer as closing the string, causing the next one to be
interpreted as an actual parenthesis, which would then be erroneously
matched against the opening brace of the function.
We fix this by removing parentheses that the semantic highlighter knows
are not actual punctuation tokens.

Change-Id: I057b9e747ed81282cdddd71a805c12db9616da65
Reviewed-by: David Schulz <david.schulz@qt.io>