Add "Files in All Project Directories" global search
Similar to the corresponding locator filter.

It searches in the project directories, not limited to the files that
are actually mentioned in the project itself. That is helpful if
projects don't mention resources, or README's, or header files, or
"alien" project files in case the project supports multiple build
systems.

In principle, for single projects, this can be achieved with the "Files
in File System" filter, but that is less convenient because one needs to
specify the directory.

Change-Id: I94bc664cadd0881274ae709b00b8576c696cc325
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>