ProjectExplorer: Fix widths distributions in project welcome page
The widths of the "Sessions" and "Projects" lists are not well
distributed on Full HD screen width. The sessions are squeezed and the
projects are extra wide.

Set a stretch factor also on the Settings widget, so that both stretch
proportionally. Use a 1 to 3 ratio.

This change also removes the setting of struts which does not appear to
have any effect and was (by mistake?) assigned to one sublayout, twice.

Fixes: QTCREATORBUG-26319
Change-Id: I365a6a6c5d9c2250e3f711cc139d3279f2f5e3b1
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>