CMake: Define BUILD_WITH_PCH before usage
On Windows QtCreatorAPIInternal.cmake would contain a check for
BUILD_WITH_PCH and if not set add a few defines to the DEFAULT_DEFINES.

First time the BUILD_WITH_PCH would not be set, because the option
will be set later, the DEFAULT_DEFINES will contain the specific non-pch
defines.

The second time BUILD_WITH_PCH would be ON and those non-pch
specific DEFAULT_DEFINES will be removed, and the build system has
to do a full rebuild because the defines changed.

Change-Id: I3f039a91667affc35f18103cfed062481f9dc93e
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>