QmlDesigner: Inject QtQuick.Controls into possible imports if necessary
In Qt6, QtQuick.Controls module itself doesn't contain any versioned
types, so it will only appear as unversioned module in the QmlJS
snaphot. It is still valid import, as behind the scenes it will import
QtQuick.Controls.impl, which has the versioned types. However, QmlJS
can't handle this case, so we work around it by including
QtQuick.Controls as a possible import if corresponding .impl is
a possible import.

Fixes: QDS-5182
Change-Id: I6cc73ccaee0c16f0bcd09e4279cba411a76b45f2
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>