Docker: Remove DockerRunConfiguration
This was a helper in the initial work on docker support which has
been commented out for a while already.

Using "docker runconfig" would technically work, however, this is
conceptionally the same as the old "matrix problem": We'd need
docker-specific almost-duplicates for every existing non-docker
runconfig. This does not scale.

Change-Id: Idf96abb6487903fdb89e7168a5ed78b9dc9dcecf
Reviewed-by: David Schulz <david.schulz@qt.io>