Themes: Do not force Fusion style, use macOS dark mode instead
We forced Fusion style on all platforms if the theme was a dark one, so
the dark appearance would be handled correctly even if the "native"
QStyle didn't.

On macOS, use the system's dark mode instead, which handles window
decoration correctly and uses native controls. Switching the QStyle also
messed up things like labels of standard buttons.

Remove the hardcoded "Fusion" style from all themes and set the
preferred style to Fusion automatically on Windows and Linux if the
theme states a DarkUserInterface and no PreferredStyles is set.

Fixes: QTCREATORBUG-22477
Change-Id: I91c9143a8703fcec7aa08201de9fc33d1799196d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>