QmlDesigner: Adjust ComponentTextModifier offsets for whitespace change
Whitespace change can also occur inside the subcomponent code when
new nodes are added into subcomponent and the document is saved.
Changed the text change handling to adjust for whitespace changes both
before and inside the subcomponent code.

The new handling is simpler and doesn't require subcomponent to have
id to work. The downside is that non-whitespace changes before
subcomponent code are no longer adjusted for, but that was just a
hacky workaround for a small part of a much larger issue that needs
to be tackled properly anyway (i.e. supporting arbitrary
undo/redo/manual edits), so losing it is not a big deal.

Fixes: QDS-5362
Change-Id: I1ca1084baf78709727b1b4726b4c034558063401
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>