QmlDesigner: Handle undo/redo/manual edits while editing subcomponent
Undo/redo stack is reset when changing between subcomponent and
document edit modes, so we can assume any undo/redo is contained within
the component. Manual edits can still happen outside the component
code, but we can assume they are always limited to either outside or
inside the component code, as the only way to make a manual edit
that spans both is to delete a block that includes both, which
invalidates the subcomponent anyway.

Fixes: QDS-5392
Change-Id: I820de73f519215a9cb3672abc55d5aa60cce078a
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>