Update background colors in all themes other than design dark
While we have chosen lower contrast lighter values for the design dark
theme (via normalBackground and alternateBackground from Palette), the
New Project dialog box did not look well on other themes. I have
replaced the background colors for those other themes, so that
DSBackgroundColorNormal is now set to the value of
DSsectionHeadBackground, while DSBackgroundColorAlternate is set to the
value of DSpanelBackground, which makes the dialog look more like the
property editor.

Also, fixed the item delegate in NewProjectView, which had a
non-transparent background. This background was visible on light
versions of the theme.

Change-Id: Iad7dea1ea1a03839293148178dbc441c05b3fb7f
Reviewed-by: Brook Cronin <brook.cronin@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>