QDS NewProject dialog: use friendly error message when regex does not match
Instead of telling the user that the project name does not match a
complex regular expression, we can tell him in English words what the
rules are.

Also, updates the icon of the status message so that, when the status is
multiline, the icon appears at the top, rather than vertical center, on
the left side of the text.

Change-Id: I848b3858f5f92a09588b9646c56cbf743b06ea09
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>