qmake/qbs/CMake builds: Remove "recommended" dependency on Designer
Recommendation of Designer is the artifact from an earlier solution
for adding code of .ui files to the code model. The refactoring of
extra compilers made that superfluous, this change removes the
left-over traces.

The CMake build of Qt Creator breaks because of the recommendation
if the qttols module is not installed.

Amends: 78d3191363883db7023a5933a18b29c716b73d7d

Change-Id: I64827492d79a9238dc8512047abd581f9da3189e
Reviewed-by: Eike Ziller <eike.ziller@qt.io>