macOS: Fix crash when quickly canceling device test dialog
and the settings dialog, by quickly pressing escape (twice) after
closing the device setup dialog.

The device test dialog was created on the stack, and run with exec(),
which creates another nested event loop for it. If the escape keys got
into the event queue before the device test dialog was shown, the first
key event was handled by the device test dialog correctly, but the
second key event was delivered to the settings dialog _within the device
test dialog's event loop_.

So, while the stack still was within DeviceSettingsWidget::testDevice()
and the dlg.exec() call, the settings dialog was rejected, and
recursively deleting children, including the device settings widget and
even worse the device test dialog - which was created on the stack. That
leads to a crash.

Move the device test dialog to the heap and use show() instead of
exec(), which fixes the issue and is the preferred method anyhow.

Fixes: QTCREATORBUG-26285
Change-Id: I03a0d1b92e466e18b753cad11482afdae2da24d0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>