QtSupport: Set a lower single step in ExamplesPageWidget view
By default, the view scrolls by a whole row when using arrow keys or
mouse wheel. With rows with a height as in the examples view, this
causes more of a fast flickering than the impression of a scrolling.

The single step value of 25 which is set by this change makes it look
smoother.

Change-Id: I7fe9fe49dab71cebdfd8b01eef245813783426ae
Reviewed-by: Christian Stenger <christian.stenger@qt.io>