GitHub Actions: Store ccache storage directory as artifact
GitHub Actions cache action uses the branch name to scope the file
caching, which when using with a gerrit bot can lead to cache
misses.

By using the cache storage as an artifact we can always get the
artifacts if previous jobs created them.

By upgrading to v2 of actions/download/upload artifact we remove
the v1 problem of the upload action that creates a directory name
with the name of the artifact, which leads to failed release
uploads.

Change-Id: I8cb8c3242b58f709f51202852ad45d42b833fa84
Reviewed-by: Eike Ziller <eike.ziller@qt.io>