Fix icon for headers and sources on macOS/Windows
On macOS and Windows, we used the system icon for C/C++ source and
header files. This was probably done under the assumption that the
system would have sensible icons registered for these types via Xcode/
Visual Studio, but actually that assumption is not very well founded.

For example I have set emacs as the default editor when double clicking
C/C++ files, and then get the same little emacs icon on all .h and .cpp
files in Qt Creator, which is not useful. There seem to be more effects
on macOS 12 leading to a generic icon too.

Fixes: QTCREATORBUG-26586
Change-Id: I88616b28d51f1583324bac9c802c9cc5bbc8ee70
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>