CppEditor: Run the built-in "local uses" code even with clangd enabled
The built-in code model treats the "local uses" result state as an
indicator of whether to provide quickfix operations. It's unclear what
kind of side effects would result from removing this condition, so we
always run the built-in variant of the "local uses" code, and only
refrain from actually updating the selections in the clangd case.
Amends 0c53c2daef.

Fixes: QTCREATORBUG-26649
Change-Id: I9e1a3e3ad08bf13bcb35efe996a61aa34302b166
Reviewed-by: David Schulz <david.schulz@qt.io>